Added proper checks for TotallyWholesome being loaded and ready, added local intensity and duration limits
